move data modules to mesmer namespace
