+++ docs/source/api.rst
@@ -110,7 +110,7 @@ Example and test data
 .. autosummary::
    :toctree: generated/
 
-   ~core.example_data.cmip6_ng_path
+   ~example_data.cmip6_ng_path
 
 Grid manipulation
 -----------------
@@ -118,12 +118,12 @@ Grid manipulation
 .. autosummary::
    :toctree: generated/
 
-   ~core.grid.wrap_to_180
-   ~core.grid.wrap_to_360
-   ~core.grid.stack_lat_lon
-   ~core.grid.unstack_lat_lon_and_align
-   ~core.grid.unstack_lat_lon
-   ~core.grid.align_to_coords
+   ~grid.wrap_to_180
+   ~grid.wrap_to_360
+   ~grid.stack_lat_lon
+   ~grid.unstack_lat_lon_and_align
+   ~grid.unstack_lat_lon
+   ~grid.align_to_coords
 
 Masking regions
 ---------------
@@ -131,9 +131,9 @@ Masking regions
 .. autosummary::
    :toctree: generated/
 
-   ~core.mask.mask_ocean_fraction
-   ~core.mask.mask_ocean
-   ~core.mask.mask_antarctica
+   ~mask.mask_ocean_fraction
+   ~mask.mask_ocean
+   ~mask.mask_antarctica
 
 Weighted operations: calculate global mean
 ------------------------------------------
@@ -141,10 +141,10 @@ Weighted operations: calculate global mean
 .. autosummary::
    :toctree: generated/
 
-   ~core.weighted.global_mean
-   ~core.weighted.lat_weights
-   ~core.weighted.weighted_mean
-   ~core.weighted.equal_scenario_weights_from_datatree
+   ~weighted.global_mean
+   ~weighted.lat_weights
+   ~weighted.weighted_mean
+   ~weighted.equal_scenario_weights_from_datatree
 
 Geospatial
 ----------
@@ -152,7 +152,7 @@ Geospatial
 .. autosummary::
    :toctree: generated/
 
-   ~core.geospatial.geodist_exact
+   ~geospatial.geodist_exact
 
 
 Anomalies
@@ -161,7 +161,7 @@ Anomalies
 .. autosummary::
    :toctree: generated/
 
-   ~core.anomaly.calc_anomaly
+   ~anomaly.calc_anomaly
 
 
 Emulator functions
@@ -173,5 +173,5 @@ Volcanic influence
 .. autosummary::
    :toctree: generated/
 
-   ~core.volc.fit_volcanic_influence
-   ~core.volc.superimpose_volcanic_influence
+   ~volc.fit_volcanic_influence
+   ~volc.superimpose_volcanic_influence
