+++ mesmer/mesmer_x/_conditional_distribution.py
@@ -8,15 +8,15 @@ from typing import Literal
 import numpy as np
 import xarray as xr
 
-from mesmer.core.geospatial import geodist_exact
 from mesmer.core.utils import _check_dataarray_form, _check_dataset_form
-from mesmer.core.weighted import weighted_median
+from mesmer.geospatial import geodist_exact
 from mesmer.mesmer_x import _distrib_checks, _optimizers
 from mesmer.mesmer_x._expression import Expression
 from mesmer.mesmer_x._first_guess import _FirstGuess
 from mesmer.mesmer_x._optimizers import MinimizeOptions, OptimizerNLL
 from mesmer.mesmer_x._utils import _ignore_warnings
 from mesmer.stats import gaspari_cohn
+from mesmer.weighted import weighted_median
 
 
 class ConditionalDistribution:
