+++ mesmer/mesmer_x/_probability_integral_transform.py
@@ -6,7 +6,7 @@
 import numpy as np
 import xarray as xr
 
-from mesmer.core.datatree import _datatree_wrapper
+from mesmer.datatree import _datatree_wrapper
 from mesmer.mesmer_x._conditional_distribution import ConditionalDistribution
