+++ mesmer/stats/_linear_regression.py
@@ -1,12 +1,12 @@
 import numpy as np
 import xarray as xr
 
-from mesmer.core.datatree import _datatree_wrapper
 from mesmer.core.utils import (
     _check_dataarray_form,
     _check_dataset_form,
     _to_set,
 )
+from mesmer.datatree import _datatree_wrapper
 
 
 class LinearRegression:
