+++ mesmer/stats/_smoothing.py
@@ -3,8 +3,8 @@ from typing import TypeVar
 import numpy as np
 import xarray as xr
 
-from mesmer.core.datatree import _datatree_wrapper
 from mesmer.core.utils import _check_dataarray_form
+from mesmer.datatree import _datatree_wrapper
 
 T_Xarray = TypeVar("T_Xarray", "xr.DataArray", "xr.Dataset")
