+++ tests/integration/test_calibrate_mesmer_newcodepath.py
@@ -3,7 +3,7 @@ import xarray as xr
 from filefisher import FileFinder
 
 import mesmer
-from mesmer.core.datatree import map_over_datasets
+from mesmer.datatree import map_over_datasets
 
 
 @pytest.mark.filterwarnings("ignore:No local minimum found")
