+++ tests/integration/test_draw_realisations_mesmer_newcodepath.py
@@ -3,7 +3,7 @@ import xarray as xr
 from filefisher import FileFinder
 
 import mesmer
-from mesmer.core.datatree import map_over_datasets
+from mesmer.datatree import map_over_datasets
 
 
 def load_and_prepare_forcing_data(scenarios, use_hfds):
